refined coverage-valued response
image is just a special case, explained the options available already which should particularly benefit non-image results, like DEM, weather data, etc. Generally, added reference to OGC CIS for more details.

Also, removed "description of" which was used in some elements, but not all. This should be confusing: what is the difference between daominSet and the domainSet? in both cases, some (XML or JSON or RDF or...) encoded domainSet "Specification" is returned.